requirements.txt
torch>=2.0
torchvision>=0.15

diffusers>=0.27
transformers>=4.30
accelerate>=0.28
peft>=0.10
safetensors>=0.4
huggingface-hub>=0.20

numpy>=1.24
scipy>=1.10
pillow>=10.0
opencv-python>=4.8
tqdm>=4.65

lpips>=0.1.4
pytorch-fid>=0.3
scikit-image>=0.22

datasets>=2.14
pyyaml>=6.0